Android: Refactored how search works in the app
Search tab was empty most of the times, so I've made it into a different
screen that can be triggered when the user hits the search icon on the
main app screen. the user can also perform more searchs from within the
search screen to be able to search for more games.