Refactoring saving games (#29)
* Android: Fixed some UI bug where the title wasn't bold

* Android: Cleanup

* Android + Core - Core now drives a view state and views just react to the view state

* Android: More work done to add image section to game details

The images are not mixed together and the grid is adaptive.
However, more polish is needed to get this to work properly

* Android: Added the media gallery composable

The media gallery composable is used to show a bunch of images

* Android + Core - Saving game now works (Kind of)

Saving games to library now works. Next step would be to ask for the
user's input on the platforms he/she owns and the list they want to
add the game to.

* Android: Added a preview of the library to the library screen

This is to allow me to debug easily. We can add more features around
this later.

* Android: Library screen not lets you see details when tapped.

* Android: Basic implementation of how games are saved.

The UI is shit !

* Android: Asking for platforms when saving games

* Android: Added flow layout to the platform selection chips

* Android: Added the ability to save games and specify the platform
and also the status.

The list type is calculated based on release date. If the release
date is higher than now then it's automatically wishlisted.

* Android + Core - Finished saving of games to library

* Android: Refactored the chip into it's own composable and replaced usages everywhere

* Android: removed unused code