Adding game details + Library to iOS (#47)
* iOS + Core : First working implementation of game details on iOS

Basic header is working

* iOS - More work done on the game detail page

Game detail page now has an expandable summary

* iOS: The text jumping bug has been fixed here.

The key was to make the about text .infinity length and change the
alignment

* iOS: New design for game detail screen

Working on a new design. it's been a hard day

* iOS + Core: Added a new view for game stats

* iOS: Fixing the sizing issue.

The key was to have each container view a fixed height so that things
don't try to jump when things change

* iOS : Added a full screen schreenshot viewer

User is able to now tap an image and then get a full screen preview

* iOS: Full screen screenshot viewer is now a modal

* iOS: Trailers are now show on game details page and open youtube.

* iOS: Doing some refactoring

* iOS: DLC's are now shown to the user

* iOS : Refactored the downloadable content view

* iOS: Added a addgameview to the game

This view is purely UX and tommorrow we hook it up

* iOS + Core: Adding game to library now works

* iOS + Core - Added library filters and ability to see games in library

With this feature, iOS and Android are now feature equivalent, with iOS
having more fineese. Android needs to catch up

* Core - Detekt fixes