Core - Fixing authentication bug
The AWS authetication lamba currently provides the expiration date
instead of the seconds to expiration. We were currently adding our
current time (thinking that we were getting seconds to expiration) so
that caused an issue where the expiration date on the database was set
to a riduculously high number and authentication failed to roll over to
the new token when we generated one.